gpsd_json.c: Fix 64-bit time_t on 32-bit arch warnings.

+++ gpsd_json.c
@@ -694,7 +694,7 @@ void json_subframe_dump(const struct gps_data_t *datap,
 		str_appendf(buf, buflen,
 			",\"IONO\":{\"a0\":%.5g,\"a1\":%.5g,\"a2\":%.5g,"
 			"\"a3\":%.5g,\"b0\":%.5g,\"b1\":%.5g,\"b2\":%.5g,"
-			"\"b3\":%.5g,\"A1\":%.11e,\"A0\":%.11e,\"tot\":%ld,"
+			"\"b3\":%.5g,\"A1\":%.11e,\"A0\":%.11e,\"tot\":%lld,"
 			"\"WNt\":%u,\"ls\":%d,\"WNlsf\":%u,\"DN\":%u,"
 			"\"lsf\":%d}",
 			    subframe->sub4_18.d_alpha0,
@@ -707,7 +707,7 @@ void json_subframe_dump(const struct gps_data_t *datap,
 			    subframe->sub4_18.d_beta3,
 			    subframe->sub4_18.d_A1,
 			    subframe->sub4_18.d_A0,
-			    subframe->sub4_18.t_tot,
+			    (long long)subframe->sub4_18.t_tot,
 			    (unsigned int)subframe->sub4_18.WNt,
 			    (int)subframe->sub4_18.leap,
 			    (unsigned int)subframe->sub4_18.WNlsf,
@@ -1244,11 +1244,11 @@ void json_rtcm3_dump(const struct rtcm3_t *rtcm,
 
     case 1009:
 	str_appendf(buf, buflen,
-		       "\"station_id\":%u,\"tow\":%d,\"sync\":\"%s\","
+		       "\"station_id\":%u,\"tow\":%lld,\"sync\":\"%s\","
 		       "\"smoothing\":\"%s\",\"interval\":\"%u\","
 		       "\"satcount\":\"%u\",",
 		       rtcm->rtcmtypes.rtcm3_1009.header.station_id,
-		       (int)rtcm->rtcmtypes.rtcm3_1009.header.tow,
+		       (long long)rtcm->rtcmtypes.rtcm3_1009.header.tow,
 		       JSON_BOOL(rtcm->rtcmtypes.rtcm3_1009.header.sync),
 		       JSON_BOOL(rtcm->rtcmtypes.rtcm3_1009.header.smoothing),
 		       rtcm->rtcmtypes.rtcm3_1009.header.interval,
